fix end of transmission msg
seq num 0 was has previously been designated to signal end of transmission. Problem was, with current design, where the sender doesnt wait for timeout or missing ack to re-send a msg, multiple pkts with seq n. 0 have been sent, which in the environment where multiple msgs have to be sent causes problems: when the reader tries to read the next msg, it will receive the end of transmission msg before even starting to read!

PROPOSED SOLUTION: add msg body indicating wether the msg is a command (CMD) or the body of the msg (BDY).